+++ Android/app/src/main/java/com/dergoogler/hentai/MainActivity.java
@@ -4,6 +4,7 @@ import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.annotation.SuppressLint;
+import android.app.DownloadManager;
 import android.content.ClipData;
 import android.content.ClipboardManager;
 import android.content.Context;
@@ -13,9 +14,11 @@ import android.net.Uri;
 import android.net.http.SslError;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Environment;
 import android.speech.tts.TextToSpeech;
 import android.view.KeyEvent;
 import android.view.View;
+import android.webkit.DownloadListener;
 import android.webkit.JavascriptInterface;
 import android.webkit.SslErrorHandler;
 import android.webkit.WebView;
@@ -25,10 +28,12 @@ import android.widget.Toast;
 import java.security.GeneralSecurityException;
 import java.util.Objects;
 
-
 public class MainActivity extends AppCompatActivity {
     WebView webView;
-    String urlCore = "file:///android_asset/www/index.html";
+    String langSetting = "?lang=";
+    String urlCore = "https://dergoogler.com/hentai-web/" + langSetting;
+    String urlCore_ = "192.168.178.81:5500" + langSetting; // For debugging
+    String mainURL = urlCore; // Main url
 
     @RequiresApi(api = Build.VERSION_CODES.M)
     @SuppressLint("SetJavaScriptEnabled")
@@ -37,7 +42,7 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        getWindow().getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR);
+        // getWindow().getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR);
         getWindow().setStatusBarColor(0xFF4A148C);
 
         Objects.requireNonNull(getSupportActionBar()).hide();
@@ -48,6 +53,20 @@ public class MainActivity extends AppCompatActivity {
         webView.getSettings().setSupportZoom(false);
         webView.getSettings().setDomStorageEnabled(true);
         webView.setWebViewClient(new Client());
+        webView.getSettings().setUserAgentString("HENTAI_WEB_AGENT");
+        webView.setDownloadListener((url, userAgent, contentDisposition, mimetype, contentLength) -> {
+            DownloadManager.Request request = new DownloadManager.Request(
+                    Uri.parse(url));
+
+            request.allowScanningByMediaScanner();
+            request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED); // Notify client once download is completed!
+            request.setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, "Hentai-Web");
+            DownloadManager dm = (DownloadManager) getSystemService(DOWNLOAD_SERVICE);
+            dm.enqueue(request);
+            Toast.makeText(getApplicationContext(), "Downloading File", // To notify the Client that the file is being downloaded
+                    Toast.LENGTH_LONG).show();
+
+        });
         webView.addJavascriptInterface(new Object() {
             @JavascriptInterface
             public void showMessage(String content) {
@@ -72,20 +91,38 @@ public class MainActivity extends AppCompatActivity {
             }
 
             @JavascriptInterface
+            public String BuildMANUFACTURER() {
+                return Build.MANUFACTURER;
+            }
+
+            @JavascriptInterface
+            public String BuildMODEL() {
+                return Build.MODEL;
+            }
+
+            @JavascriptInterface
+            public void reload() {
+                Intent i = new Intent(MainActivity.this, MainActivity.class);
+                finish();
+                overridePendingTransition(0, 0);
+                startActivity(i);
+                overridePendingTransition(0, 0);
+            }
+
+            @JavascriptInterface
             public void copyToClipboard(String content) {
                 ClipboardManager clipboard = (ClipboardManager) getSystemService(Context.CLIPBOARD_SERVICE);
                 ClipData clip = ClipData.newPlainText("copy", content);
                 clipboard.setPrimaryClip(clip);
             }
         }, "Android");
-
-        webView.loadUrl(urlCore);
+        webView.loadUrl(mainURL + getResources().getString(R.string.lang));
     }
 
     public class Client extends WebViewClient {
         @Override
         public boolean shouldOverrideUrlLoading(WebView view, String url) {
-            if (url.contains(urlCore)) {
+            if (url.contains(mainURL + getResources().getString(R.string.lang)) || url.contains(mainURL)) {
                 view.loadUrl(url);
             } else {
                 Intent i = new Intent(Intent.ACTION_VIEW, Uri.parse(url));