+++ Website/src/misc/strings.ts
@@ -1,9 +1,54 @@
-const string = {
-  dialog_message: "This web app contains +18 content",
-  dialog_title: "Content",
-  ok: "Ok",
-  repo_link: "https://github.com/DerGoogler/Hentai-Web",
-  ie_text: "IE is not supported. Download Chrome/Opera/Firefox",
-};
+/**
+ * Strings of the app
+ */
+export default {
+  "dialog-message": {
+    en: "This web app contains +18 content",
+    de: "Die Seite beinhalten Bilder die nicht unter 18 sind!",
+  },
 
-export default string;
+  "dialog-title": {
+    en: "Content",
+    de: "Inhalt",
+  },
+
+  ok: {
+    en: "ok",
+    de: "ok",
+  },
+
+  "download-text": {
+    en: "Download",
+    de: "Herunterladen",
+  },
+
+  "view-image": {
+    en: "View image in browser",
+    de: "Im Browser ansehen",
+  },
+
+  "repo-link": {
+    en: "https://github.com/DerGoogler/Hentai-Web",
+    de: "https://github.com/DerGoogler/Hentai-Web",
+  },
+
+  "ie-text": {
+    en: "IE is not supported. Download Chrome/Opera/Firefox",
+    de: "IE wird nicht unterstüzt. Lade Chrome/Opera/Firefox herunter",
+  },
+
+  "sign-in": {
+    en: "Sign in",
+    de: "Anmelden",
+  },
+
+  "how-to-login": {
+    en: "How to login?",
+    de: "Wie melde ich mich an?",
+  },
+
+  "how-to-login-link": {
+    en: "https://dergoogler.com/?/guide/hentai-web/en/how-to-login/",
+    de: "https://dergoogler.com/?/guide/hentai-web/de/how-to-login/",
+  },
+};