+++ Website/src/views/NSFW/data.ts
@@ -1,0 +1,123 @@
+import { hot } from "react-hot-loader/root";
+import hmtai from "hmtai";
+
+const data = [
+  {
+    name: "ass",
+    source: hmtai.nsfw.ass(),
+  },
+  {
+    name: "bdsm",
+    source: hmtai.nsfw.bdsm(),
+  },
+  {
+    name: "cum",
+    source: hmtai.nsfw.cum(),
+  },
+  {
+    name: "creampie",
+    source: hmtai.nsfw.creampie(),
+  },
+  {
+    name: "manga",
+    source: hmtai.nsfw.manga(),
+  },
+  {
+    name: "femdom",
+    source: hmtai.nsfw.femdom(),
+  },
+  {
+    name: "hentai",
+    source: hmtai.nsfw.hentai(),
+  },
+  {
+    name: "incest",
+    source: hmtai.nsfw.incest(),
+  },
+  {
+    name: "masturbation",
+    source: hmtai.nsfw.masturbation(),
+  },
+  {
+    name: "public",
+    source: hmtai.nsfw.public(),
+  },
+  {
+    name: "ero",
+    source: hmtai.nsfw.ero(),
+  },
+  {
+    name: "orgy",
+    source: hmtai.nsfw.orgy(),
+  },
+  {
+    name: "elves",
+    source: hmtai.nsfw.elves(),
+  },
+  {
+    name: "yuri",
+    source: hmtai.nsfw.yuri(),
+  },
+  {
+    name: "pantsu",
+    source: hmtai.nsfw.pantsu(),
+  },
+  {
+    name: "glasses",
+    source: hmtai.nsfw.glasses(),
+  },
+  {
+    name: "cuckold",
+    source: hmtai.nsfw.cuckold(),
+  },
+  {
+    name: "blowjob",
+    source: hmtai.nsfw.blowjob(),
+  },
+  {
+    name: "boobjob",
+    source: hmtai.nsfw.boobjob(),
+  },
+  {
+    name: "foot",
+    source: hmtai.nsfw.foot(),
+  },
+  {
+    name: "thighs",
+    source: hmtai.nsfw.thighs(),
+  },
+  {
+    name: "ahegao",
+    source: hmtai.nsfw.ahegao(),
+  },
+  {
+    name: "uniform",
+    source: hmtai.nsfw.uniform(),
+  },
+  {
+    name: "gang Bang",
+    source: hmtai.nsfw.gangbang(),
+  },
+  {
+    name: "tentacles",
+    source: hmtai.nsfw.tentacles(),
+  },
+  {
+    name: "gif",
+    source: hmtai.nsfw.gif(),
+  },
+  {
+    name: "nSFW Neko",
+    source: hmtai.nsfw.nsfwNeko(),
+  },
+  {
+    name: "nSFW Mobile Wallpaper",
+    source: hmtai.nsfw.nsfwMobileWallpaper(),
+  },
+  {
+    name: "zettai Ryouiki",
+    source: hmtai.nsfw.zettaiRyouiki(),
+  },
+];
+
+export default hot(data);