+++ Website/src/views/SFW/index.tsx
@@ -1,0 +1,21 @@
+import AnimePicture from "../../builders/AniemePicture";
+import hmtai from "hmtai";
+import * as React from "react";
+import { hot } from "react-hot-loader/root";
+import { List } from "react-onsenui";
+import data from "./data";
+
+class SFW extends React.Component {
+  render() {
+    /**
+     * To load for every object an own AnimePicture from `data.ts`
+     */
+    const listItems = data.map((item) => (
+      <AnimePicture key={item.name} source={item.source} note={item.name} />
+    ));
+
+    return <List>{listItems}</List>;
+  }
+}
+
+export default hot(SFW);